+++ example/src/App/MainRouter.tsx
@@ -9,6 +9,7 @@ import Loop from "views/Feature/Loop";
 import VipLevelList from "views/Example/VipLevelList";
 import AutoPlayProgress from "views/Example/AutoPlayProgress";
 import TextAnimations from "views/Example/TextAnimations";
+import PreviewImage from 'views/Example/PreviewImage';
 
 
 const MainRouter = () => {
@@ -20,6 +21,7 @@ const MainRouter = () => {
                 <Route path="example/vip-level-list" element={<VipLevelList/>} />
                 <Route path="example/auto-play-progress" element={<AutoPlayProgress/>} />
                 <Route path="example/text-animations" element={<TextAnimations/>} />
+                <Route path="example/preview-image" element={<PreviewImage/>} />
 
                 <Route path="" element={<BaseUsed/>} />
             </Route>